used system clock to guarantee correct timing for DataLogger and Voltmeter

+++ examples/runDataLogger.py
@@ -3,6 +3,9 @@
 
 '''Data Logger 
      reads samples from PicoScope and display averages as voltage history
+
+     Usage: ./runDataLogger.py [<Oscilloscpope_config>.yaml Interval]
+
 '''
 
 from __future__ import print_function, division, unicode_literals
@@ -20,9 +23,10 @@ def stop_processes(proclst):
     Close all running processes at end of run
   '''
   for p in proclst: # stop all sub-processes
-    print('    terminating '+p.name)
-    p.terminate()
-  time.sleep(2)
+    if p.is_alive():
+      print('    terminating '+p.name)
+      p.terminate()
+      time.sleep(1.)
 
 if __name__ == "__main__": # - - - - - - - - - - - - - - - - - - - - - -
 
@@ -30,11 +34,16 @@ if __name__ == "__main__": # - - - - - - - - - - - - - - - - - - - - - -
 
 # check for / read command line arguments
   # read DAQ configuration file
-  if len(sys.argv)==2:
+  if len(sys.argv)>=2:
     PSconfFile = sys.argv[1]
   else: 
     PSconfFile = 'PSdataLogger.yaml'
   print('    PS configuration from file ' + PSconfFile)
+
+  if len(sys.argv)==3:
+    interval = float(sys.argv[2])
+  else: 
+    interval = 0.2
 
   # read scope configuration file
   print('    Device configuration from file ' + PSconfFile)
@@ -54,9 +63,8 @@ if __name__ == "__main__": # - - - - - - - - - - - - - - - - - - - - - -
   NSamples = PSconf.NSamples   # number of samples
   buf = np.zeros( (NChannels, NSamples) ) # data buffer for PicoScope driver
 
-
   procs=[]
-  deltaT = 500.
+  deltaT = interval *1000   # minimal time between figure updates in ms
   DLmpQ =  mp.Queue(1) # Queue for data transfer to sub-process
   procs.append(mp.Process(name='DataLogger', target = mpDataLogger, 
                args=(DLmpQ, PSconf.OscConfDict, deltaT, '(Volt)') ) )
@@ -70,20 +78,21 @@ if __name__ == "__main__": # - - - - - - - - - - - - - - - - - - - - - -
 
     sig = np.zeros(NChannels)
 
-# -- LOOP 
+  # -- LOOP 
   try:
     print('          type <cntrl>C to end -->')
     while True:
-      if DLmpQ.empty(): 
-        PSconf.acquireData(buf) # read data from PicoScope
-        for i, b in enumerate(buf): # process data 
-          # sig[i] = np.sqrt (np.inner(b, b) /NSamples) # eff. Voltage
-          sig[i] = b.sum() / NSamples # average
-        DLmpQ.put(sig) 
+      PSconf.acquireData(buf) # read data from PicoScope
+      for i, b in enumerate(buf): # process data 
+        # sig[i] = np.sqrt (np.inner(b, b) / NSamples)    # eff. Voltage
+        sig[i] = b.sum() / NSamples          # average
+      DLmpQ.put(sig) 
 
   except KeyboardInterrupt:
      print(sys.argv[0]+': keyboard interrupt - closing down ...')
   finally:
     PSconf.closeDevice() # close down hardware device
+    DLmpQ.put(None) 
+    time.sleep(1.)
     stop_processes(procs)  # stop all sub-processes in list
     print('*==* ' + sys.argv[0] + ': normal end')
